Refactored the VAE examples
- Created an abstract VAE class that can be implemented for specific probability distributions.
- Implemented the code for normal and categorical distributions
