+++ examples/vae/__init__.py
@@ -1,0 +1,2 @@
+from examples.vae.train import main
+from examples.vae.vae import VAE
